11.05.2017 Updated base.py, node.py
Completed the secondary sequence for single transmission style and added timing as well as transmission of the times back to the base station. Added modes to base.py, one that samples the time between the base station and each node in two sets of data and then continues sampling until the peaks of the Gaussian kernel density equations are within a specified threshold (until they have converged) and a second mode where the time between the base station and each node a fixed number of times.